LayoutTests/fast/viewport/viewport-83-expected.txt

CONSOLE MESSAGE: line 3: Viewport argument value "height" for key "width" not recognized. Content ignored.
CONSOLE MESSAGE: line 3: Viewport argument key "830" not recognized and ignored.
ALERT: viewport size 64x70.4 scale 5 with limits [5, 5] and userScalable true
